+++ CHANGELOG.md
@@ -46,7 +46,7 @@
 - [fix(useSelect): fixed clear button on empty select  (#1473)](https://github.com/gazprom-neft/consta-uikit/commit/b918af45477a969b0a6b0273c1b8de414298ccfb) - [@N1KN](https://github.com/N1KN)
 - [feat(RadioGroup): add new property for RadioGroup (align) (#1465)](https://github.com/gazprom-neft/consta-uikit/commit/ce4efd071dd761d8d0da4bbe3746f2585fecc717) - [@N1MBER](https://github.com/N1MBER)
 - [fix(cn fuction): fix bug with cn fuction](https://github.com/gazprom-neft/consta-uikit/commit/8c33625c46810cb4170b15d40620568e339157ca) - [@N1MBER](https://github.com/N1MBER)
-- [refactor(SnackBar): change description of props types](https://github.com/gazprom-neft/consta-uikit/commit/4f486dc1aca18219f986e9d49c2b4e6b90286666) - [@N1MBER](https://github.com/N1MBER)
+- [refactor(Notification): change description of props types](https://github.com/gazprom-neft/consta-uikit/commit/4f486dc1aca18219f986e9d49c2b4e6b90286666) - [@N1MBER](https://github.com/N1MBER)
 - [refactor(Tooltip): delete preview container from tooltip documentation](https://github.com/gazprom-neft/consta-uikit/commit/4651448dadd54721f7528ab91a4da022856925df) - [@N1MBER](https://github.com/N1MBER)
 
 --------------------